requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.3
peewee==3.14.10
python-telegram-bot==13.11
python-telegram-bot-pagination==0.0.2
requests==2.27.1
matplotlib==3.5.1